Implement 500 msg buffering toward logstash upon conn failure.
If the connection to logstash (TCP) is lost, buffer messages which arrive
until we have connectivity. Then once connectivity is established, drain
the buffer payload on the TCP socket.

If more than 500 msgs are in the buffer. Kill the buffer and start all over.
This ensures proper drainage if something goes awry.